backend/requirements.txt
# FastAPI
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# AI/ML - Core packages
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.4.0

# Computer Vision
opencv-python-headless>=4.8.0
scikit-image>=0.20.0
scipy>=1.10.0
shapely>=2.0.0
pillow>=10.0.0

# Segment Anything
git+https://github.com/facebookresearch/segment-anything.git

# Database & Storage
supabase>=2.0.0
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
httpx>=0.24.0
